Tonuino: Modifier modifications*. Extent RFID tool to support new modifiers.
*: - Add modifier to play advertisement
    - Change modifier value type to uint16
    - Combine track play, pause and toggle modifiers to TrackContinue

+++ Tonuino_Main/src/shared/Tonuino_DFPlayer.cpp
@@ -257,7 +257,7 @@ void TonuinoDFPlayer::waitForTrackToFinish()
 	} while (isPlaying());
 }
 
-void TonuinoDFPlayer::playAdvertisement(int advertisement)
+void TonuinoDFPlayer::playAdvertisement(uint16_t advertisement)
 {
 	Serial.print(F("Play advertisement "));
 	Serial.println(advertisement);
